针对前端传来数据做hash判断，如果一样的，那么不落库

+++ src/main/java/com/jianli/component/Html2PdfImpl.java
@@ -3,6 +3,7 @@ package com.jianli.component;
 import com.jianli.commons.FileUtils;
 import com.jianli.exception.PdfException;
 import com.pdfcrowd.Pdfcrowd;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Component;
  * @date 2018-09-02 22:33
  */
 @Component
+@Slf4j
 public class Html2PdfImpl implements Html2Pdf {
     private final String username;
     private final String apiKey;
@@ -21,6 +23,10 @@ public class Html2PdfImpl implements Html2Pdf {
     }
     @Override
     public void writeTO(String text, String origin, String dest) {
+        if (FileUtils.exist(origin)) {
+            log.info("已经生成过的文档，不需要再次生成,origin:{}, dest:{}", origin, dest);
+            return ;
+        }
         // 写入原始HTML文件
         FileUtils.writeToWithoutAppend(text, origin);
         try {
